Clean up BrowseApp (#14217)
* move BrowseApp into own directory

* extract browse app containers, use functions instead of classes

* remove duplicate table loader, organize import
