+++ frontend/src/metabase/browse/components/BrowseApp.jsx
@@ -1,0 +1,8 @@
+import React from "react";
+import { Box } from "grid-styled";
+
+import { PAGE_PADDING } from "metabase/browse/constants";
+
+export default function BrowseApp({ children }) {
+  return <Box mx={PAGE_PADDING}>{children}</Box>;
+}
